generalize Card component and move child profile logic to ProfilePage
