+++ src/app/[locale]/profile/page.tsx
@@ -11,8 +11,15 @@ interface ChildProfile {
   allergies: string[];
 }
 
+const allergyOptions = [
+  "Peanut", "Milk", "Egg", "Soy", 
+  "Wheat", "Fish", "Shellfish", "Tree nuts", "Chicken", "Celery"
+];
+
 export default function ProfilePage() {
   const [children, setChildren] = useState<ChildProfile[]>([]);
+  const [editIndex, setEditIndex] = useState<number | null>(null);
+
   const CHILDREN_KEY = "user_children";
 
   useEffect(() => {
@@ -20,11 +27,12 @@ export default function ProfilePage() {
     if (saved) setChildren(saved);
   }, []);
 
-  const handleUpdate = (index: number, updatedChild: ChildProfile) => {
+  const handleSave = (index: number, updatedChild: ChildProfile) => {
     const updatedChildren = [...children];
     updatedChildren[index] = updatedChild;
     setChildren(updatedChildren);
     storageService.setLocalItem(CHILDREN_KEY, updatedChildren);
+    setEditIndex(null);
   };
 
   const handleClear = () => {
@@ -44,9 +52,48 @@ export default function ProfilePage() {
           {children.map((child, index) => (
             <Card
               key={index}
-              child={child}
-              onUpdate={(updated) => handleUpdate(index, updated)}
-            />
+              title={child.name}
+              actions={
+                editIndex === index ? null : (
+                  <button
+                    onClick={() => setEditIndex(index)}
+                    className="text-blue-500 text-sm hover:underline"
+                  >
+                    Edit
+                  </button>
+                )
+              }
+            >
+              {editIndex === index ? (
+                <EditForm
+                  initialData={child}
+                  onSave={(updatedChild) => handleSave(index, updatedChild)}
+                  onCancel={() => setEditIndex(null)}
+                />
+              ) : (
+                <>
+                  <p><strong>Age:</strong> {child.age} years</p>
+                  <p><strong>Gender:</strong> {child.gender}</p>
+                  <div className="mt-4">
+                    <p className="font-semibold mb-2">Allergies:</p>
+                    <div className="flex flex-wrap gap-2">
+                      {child.allergies.length > 0 ? (
+                        child.allergies.map((allergy, idx) => (
+                          <span
+                            key={idx}
+                            className="bg-green-100 text-green-800 px-3 py-1 rounded-full text-sm font-medium"
+                          >
+                            {allergy}
+                          </span>
+                        ))
+                      ) : (
+                        <p className="text-gray-500 text-sm">No allergies recorded</p>
+                      )}
+                    </div>
+                  </div>
+                </>
+              )}
+            </Card>
           ))}
         </div>
       ) : (
@@ -59,6 +106,113 @@ export default function ProfilePage() {
       >
         Clear All Profiles
       </button>
+    </div>
+  );
+}
+
+function EditForm({
+  initialData,
+  onSave,
+  onCancel
+}: {
+  initialData: ChildProfile;
+  onSave: (updated: ChildProfile) => void;
+  onCancel: () => void;
+}) {
+  const [name, setName] = useState(initialData.name);
+  const [age, setAge] = useState(initialData.age);
+  const [gender, setGender] = useState(initialData.gender);
+  const [allergies, setAllergies] = useState(initialData.allergies);
+  const [otherAllergy, setOtherAllergy] = useState('');
+
+  const toggleAllergy = (item: string) => {
+    if (allergies.includes(item)) {
+      setAllergies(allergies.filter((a) => a !== item));
+    } else {
+      setAllergies([...allergies, item]);
+    }
+  };
+
+  const handleSave = () => {
+    let finalAllergies = [...allergies];
+    if (otherAllergy.trim()) finalAllergies.push(otherAllergy.trim());
+
+    onSave({
+      name,
+      age,
+      gender,
+      allergies: finalAllergies
+    });
+  };
+
+  return (
+    <div className="flex flex-col space-y-3">
+      <input
+        value={name}
+        onChange={(e) => setName(e.target.value)}
+        className="w-full border rounded px-3 py-2"
+        placeholder="Name"
+      />
+      <select
+        value={age}
+        onChange={(e) => setAge(e.target.value)}
+        className="w-full border rounded px-3 py-2"
+      >
+        <option value="">Select Age</option>
+        {[5,6,7,8,9,10,11,12].map(num => (
+          <option key={num} value={num}>{num} years</option>
+        ))}
+      </select>
+      <select
+        value={gender}
+        onChange={(e) => setGender(e.target.value)}
+        className="w-full border rounded px-3 py-2"
+      >
+        <option value="">Select Gender</option>
+        <option value="male">Male</option>
+        <option value="female">Female</option>
+        <option value="other">Other</option>
+      </select>
+
+      <div>
+        <label className="font-semibold text-sm">Allergies</label>
+        <div className="flex flex-wrap gap-2 my-2">
+          {allergyOptions.map((item) => (
+            <span
+              key={item}
+              onClick={() => toggleAllergy(item)}
+              className={`cursor-pointer px-3 py-1 rounded-full text-sm font-medium ${
+                allergies.includes(item)
+                  ? 'bg-green-200 text-green-800'
+                  : 'bg-gray-200 text-gray-700'
+              }`}
+            >
+              {item}
+            </span>
+          ))}
+        </div>
+        <input
+          value={otherAllergy}
+          onChange={(e) => setOtherAllergy(e.target.value)}
+          placeholder="Other allergy"
+          className="w-full border rounded px-3 py-2 mt-2"
+        />
+      </div>
+
+      <div className="flex gap-4 mt-2">
+        <button
+          onClick={handleSave}
+          className="bg-green-500 text-white py-1 px-4 rounded hover:bg-green-600"
+        >
+          Save
+        </button>
+        <button
+          onClick={onCancel}
+          className="text-gray-600 hover:underline"
+        >
+          Cancel
+        </button>
+      </div>
     </div>
   );
 }
